LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/text/2d.text.draw.baseline.bottom-expected.txt

2d.text.draw.baseline.bottom
textBaseline bottom is the bottom of the em square (not the bounding box)
Actual output:
Expected output:


FAIL textBaseline bottom is the bottom of the em square (not the bounding box) assert_approx_equals: Red channel of the pixel at (25, 25) expected 0 +/- 2 but got 255
